Refactoring print_json_hist (#3211)
implement print_json_hist in table.py as a common API.
refactor print_log2_hist and print_linear_hist to abstract out the common part and
reuse it for print_json_hist.
